+++ src/elements/Category.php
@@ -616,7 +616,7 @@ class Category extends Element
         if ($this->structureId) {
             // Remember the parent ID, in case the entry needs to be restored later
             $parentId = $this->getAncestors(1)
-                ->anyStatus()
+                ->status(null)
                 ->select(['elements.id'])
                 ->scalar();
             if ($parentId) {
@@ -668,7 +668,7 @@ class Category extends Element
             $newRelationValues = [];
 
             $ancestorIds = $this->getAncestors()
-                ->anyStatus()
+                ->status(null)
                 ->ids();
 
             $sources = (new Query())
@@ -757,7 +757,7 @@ class Category extends Element
         $oldParentQuery->ancestorOf($this);
         $oldParentQuery->ancestorDist(1);
         $oldParentQuery->siteId($this->siteId);
-        $oldParentQuery->anyStatus();
+        $oldParentQuery->status(null);
         $oldParentQuery->select('elements.id');
         $oldParentId = $oldParentQuery->scalar();