+++ src/elements/User.php
@@ -447,7 +447,7 @@ class User extends Element implements IdentityInterface
         $user = static::find()
             ->addSelect(['users.password'])
             ->id($id)
-            ->anyStatus()
+            ->status(null)
             ->one();
 
         if ($user === null) {
@@ -463,7 +463,7 @@ class User extends Element implements IdentityInterface
         if ($previousUserId = Session::get(self::IMPERSONATE_KEY)) {
             $previousUser = static::find()
                 ->id($previousUserId)
-                ->anyStatus()
+                ->status(null)
                 ->one();
 
             if ($previousUser && $previousUser->can('impersonateUsers')) {
@@ -807,7 +807,7 @@ class User extends Element implements IdentityInterface
     public function validateUnverifiedEmail(string $attribute, $params, InlineValidator $validator)
     {
         $query = self::find()
-            ->anyStatus();
+            ->status(null);
 
         if (Craft::$app->getDb()->getIsMysql()) {
             $query->where([
@@ -1628,7 +1628,7 @@ class User extends Element implements IdentityInterface
                     ->siteId('*')
                     ->unique()
                     ->authorId($this->id)
-                    ->anyStatus();
+                    ->status(null);
 
                 foreach (Db::each($entryQuery) as $entry) {
                     $elementsService->deleteElement($entry);