+++ src/fields/BaseRelationField.php
@@ -896,7 +896,7 @@ JS;
     {
         if ($value instanceof ElementQueryInterface) {
             $value = $value
-                ->anyStatus()
+                ->status(null)
                 ->all();
         } else if (!is_array($value)) {
             $value = [];
@@ -1070,7 +1070,7 @@ JS;
     {
         $clone = clone $query;
         $clone
-            ->anyStatus()
+            ->status(null)
             ->siteId('*')
             ->limit(null)
             ->unique();